requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.0
flake8==6.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytest==7.2.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
